pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "analysis>=0.1.7",
    "cryptography>=45.0.6",
    "feedparser>=6.0.11",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "prowler>=3.16.9",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "signals>=0.0.2",
    "streamlit>=1.48.0",
    "ta-lib-easy>=1.0.3",
    "trafilatura>=2.0.0",
    "urllib3>=2.5.0",
    "websockets>=15.0.1",
]
